% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calculate_i_squared}
\alias{calculate_i_squared}
\title{Calculate I-squared Statistic}
\usage{
calculate_i_squared(Q, df)
}
\arguments{
\item{Q}{Q statistic}

\item{df}{Degrees of freedom}
}
\value{
I-squared value
}
\description{
Calculate I-squared Statistic
}
