% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{format_p}
\alias{format_p}
\title{Format p-values for display.-values.}
\usage{
format_p(p, digits = 3)
}
\arguments{
\item{p}{P-values}

\item{digits}{Number of decimal places}
}
\value{
Formatted p-values
}
\description{
Format p-values for consistent display
}
