% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbamm_original_20250904_134106.R, R/methods.R
\name{plot.cbamm_results}
\alias{plot.cbamm_results}
\title{Plot method for cbamm_results}
\usage{
\method{plot}{cbamm_results}(x, which = "all", ...)

\method{plot}{cbamm_results}(x, which = "all", ...)
}
\arguments{
\item{x}{CBAMM results object}

\item{which}{Which plot to show}

\item{...}{Additional arguments}
}
\value{
Plot object
}
\description{
Plot method for cbamm_results

Plot Method for CBAMM Results
}
