% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_analysis.R
\name{robust_rma}
\alias{robust_rma}
\title{Robust Meta-Analysis with HKSJ Adjustment}
\usage{
robust_rma(
  yi,
  sei,
  data = NULL,
  method = "REML",
  weights = NULL,
  mods = NULL,
  use_hksj = TRUE
)
}
\arguments{
\item{yi}{Effect sizes}

\item{sei}{Standard errors}

\item{data}{Optional data frame}

\item{method}{Variance estimator method}

\item{weights}{Optional study weights}

\item{mods}{Optional moderators}

\item{use_hksj}{Whether to use Hartung-Knapp-Sidik-Jonkman adjustment}
}
\value{
Meta-analysis results object
}
\description{
Robust Meta-Analysis with HKSJ Adjustment
}
