% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/novelty.R
\name{run_adaptive_advisor}
\alias{run_adaptive_advisor}
\title{Run Adaptive Advisor}
\usage{
run_adaptive_advisor(data, pooled_results, config)
}
\arguments{
\item{data}{Study data}

\item{pooled_results}{Pooled meta-analysis results}

\item{config}{CBAMM configuration}
}
\value{
List with advisor recommendations
}
\description{
Run Adaptive Advisor
}
\keyword{internal}
