% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{safe_divide}
\alias{safe_divide}
\title{Protected division operation..}
\usage{
safe_divide(numerator, denominator, na_value = NA)
}
\arguments{
\item{numerator}{Numerator}

\item{denominator}{Denominator}

\item{na_value}{Value to return when denominator is zero}
}
\value{
Result of division
}
\description{
Division with handling for zero denominators
}
\keyword{internal}
