% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance.R
\name{should_skip_bayesian}
\alias{should_skip_bayesian}
\title{Check if Bayesian analysis should be skipped}
\usage{
should_skip_bayesian(data, min_studies = 6)
}
\arguments{
\item{data}{Input data}

\item{min_studies}{Minimum studies threshold}
}
\value{
Logical
}
\description{
Check if Bayesian analysis should be skipped
}
\keyword{internal}
