% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance.R
\name{should_skip_expensive}
\alias{should_skip_expensive}
\title{Check if expensive analyses should be skipped}
\usage{
should_skip_expensive(data, min_studies = 8)
}
\arguments{
\item{data}{Input data}

\item{min_studies}{Minimum studies threshold}
}
\value{
Logical
}
\description{
Check if expensive analyses should be skipped
}
\keyword{internal}
