% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{summarize_weights}
\alias{summarize_weights}
\title{Generate Study Weights Summary}
\usage{
summarize_weights(weights, study_ids = NULL)
}
\arguments{
\item{weights}{Study weights}

\item{study_ids}{Study identifiers}
}
\value{
Data frame with weight summary
}
\description{
Generate Study Weights Summary
}
