% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{var_to_se}
\alias{var_to_se}
\title{Compute Standard Error from Variance}
\usage{
var_to_se(vi)
}
\arguments{
\item{vi}{Variances}
}
\value{
Standard errors
}
\description{
Compute Standard Error from Variance
}
